Delete an Article - Fixed

+++ BlazingBlog.Application/Articles/DeleteArticle/DeleteArticleCommandHandler.cs
@@ -1,15 +1,24 @@
-﻿namespace BlazingBlog.Application.Articles.DeleteArticle
+﻿using BlazingBlog.Application.Users;
+
+namespace BlazingBlog.Application.Articles.DeleteArticle
 {
 	public class DeleteArticleCommandHandler : ICommandHandler<DeleteArticleCommand>
 	{
 		private readonly IArticleRepository _articleRepository;
+		private readonly IUserService _userService;
 
-		public DeleteArticleCommandHandler(IArticleRepository articleRepository)
+		public DeleteArticleCommandHandler(IArticleRepository articleRepository, IUserService userService)
 		{
 			_articleRepository = articleRepository;
+			_userService = userService;
 		}
 		public async Task<Result> Handle(DeleteArticleCommand request, CancellationToken cancellationToken)
 		{
+			if (!await _userService.CurrentUserCanEditArticleAsync(request.Id))
+			{
+				return Result.Fail<ArticleResponse?>("You're not allowed to delete this article!");
+			}
+
 			var deleted = await _articleRepository.DeleteArticleAsync(request.Id);
 			if (deleted)
 				return Result.Ok();
